+++ Scripts/public-api-diff/Sources/Helpers/SDKAnalyzer+Change.swift
@@ -1,0 +1,42 @@
+//
+//  File.swift
+//  
+//
+//  Created by Alexander Guretzki on 21/06/2024.
+//
+
+import Foundation
+
+extension SDKAnalyzer {
+    
+    struct Change {
+        enum ChangeType {
+            case addition
+            case removal
+            case change
+            
+            var icon: String {
+                switch self {
+                case .addition: "❇️ "
+                case .removal: "😶‍🌫️"
+                case .change: "🔀"
+                }
+            }
+        }
+        
+        var changeType: ChangeType
+        var parentName: String
+        var changeDescription: String
+    }
+}
+
+extension [String: [SDKAnalyzer.Change]] {
+    
+    var totalChangeCount: Int {
+        var totalChangeCount = 0
+        keys.forEach { targetName in
+            totalChangeCount += self[targetName]?.count ?? 0
+        }
+        return totalChangeCount
+    }
+}
